+++ pioreactor/types.py
@@ -6,13 +6,18 @@ import typing as t
 
 from msgspec import Meta
 
+if t.TYPE_CHECKING:
+    from pioreactor.pubsub import Client
+
 
 class DosingProgram(t.Protocol):
     """
     Should return a non-negative float representing (approx) how much liquid was moved, in ml.
     """
 
-    def __call__(self, unit: str, experiment: str, ml: float, source_of_event: str) -> float:
+    def __call__(
+        self, unit: str, experiment: str, ml: float, source_of_event: str, mqtt_client: Client
+    ) -> float:
         # don't forget to return a float!
         ...